Renaming EqualityVerifier to SingleValueVerifier
Reducing confusing with the EqualsVerifier utility
